pyproject.toml
[project]
name = "ai-ui-telegram-bot"
version = "1.0.0"
description = "Telegram bot for generating automated tests using AI agent"
authors = [
    {name = "utb404", email = "utb404@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "browser-use>=0.9.1",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
    "pytest>=7.4.0",
    "allure-pytest>=2.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["sample_ui"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--alluredir=allure-results"
